added solution Level 11; moved files in package; create utils classes.
