+++ src/by/devincubator/service/RentVehicle.java
@@ -1,6 +1,9 @@
-package by.devincubator.vehicle;
+package by.devincubator.service;
 
 import by.devincubator.exception.DefectedVehicleException;
+import by.devincubator.service.MechanicService;
+import by.devincubator.vehicle.Rent;
+import by.devincubator.vehicle.Vehicle;
 
 import java.util.Date;
