+++ src/by/devincubator/vehicle/Vehicle.java
@@ -1,6 +1,7 @@
 package by.devincubator.vehicle;
 
-import by.devincubator.engine.Startable;
+import by.devincubator.service.TechnicalSpecialist;
+import by.devincubator.vehicle.engine.Startable;
 import by.devincubator.exception.NotVehicleException;
 
 import java.util.List;
