+++ src/by/devincubator/vehicle/comparator/ComparatorByDefectCount.java
@@ -1,4 +1,7 @@
-package by.devincubator.vehicle;
+package by.devincubator.vehicle.comparator;
+
+import by.devincubator.service.MechanicService;
+import by.devincubator.vehicle.Vehicle;
 
 import java.util.Comparator;
