+++ src/by/devincubator/vehicle/engine/Startable.java
@@ -1,4 +1,4 @@
-package by.devincubator.engine;
+package by.devincubator.vehicle.engine;
 
 public interface Startable {
     double getTaxPerMonth();
